pyproject.toml
[tool.poetry]
name = "simple-langchain"
version = "0.1.0"
description = ""
authors = ["Tibor Reiss <tibor.reiss@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-cohere = "^0.1.8"
python-dotenv = "^1.0.1"
traceloop-sdk = "^0.24.0"
langchain = "^0.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
